pull common fields and methods up
